# chore(parser): Optimize string parsing